+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.0].define(version: 2022_05_22_223332) do
+ActiveRecord::Schema[7.0].define(version: 2022_05_25_034325) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "citext"
   enable_extension "plpgsql"
@@ -116,6 +116,7 @@ ActiveRecord::Schema[7.0].define(version: 2022_05_22_223332) do
     t.datetime "updated_at", null: false
     t.string "resident_status"
     t.date "verified_on"
+    t.boolean "residence"
     t.index ["property_id"], name: "index_residencies_on_property_id"
     t.index ["resident_id"], name: "index_residencies_on_resident_id"
     t.index ["resident_status"], name: "index_residencies_on_resident_status"